pyproject.toml
[tool.poetry]
name = "bank"
version = "0.1.0"
description = ""
authors = ["newbission <newbission@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.7"}
aerich = "^0.7.2"
python-dotenv = "^1.0.1"



[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./app/migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
